git-svn-id: https://svn.esynctraining.com/svn/esynctraining@13111 41667c05-fd31-8a45-9ea7-b4ae18c5af78

+++ EdugameCloud.Lti.Core/DTO/ACSessionParticipantReportDTO.cs
@@ -55,7 +55,7 @@ namespace EdugameCloud.Lti.Core.DTO
         }
         public string firstName { get; set; }
         public string lastName { get; set; }
-        public int transcriptId { get; set; }
+        public long transcriptId { get; set; }
 
         #endregion
